Integracao/ContaCapitalIntegracaoEJB/src/test/java/br/com/sicoob/sisbr/cca/integracao/negocio/servicos/interceptors/ContaCapitalIntegracaoInternacionalizacaoInterceptorTest.java
package br.com.sicoob.sisbr.cca.integracao.negocio.servicos.interceptors;

import org.junit.Assert;
import org.junit.BeforeClass;
import org.junit.Test;

import br.com.bancoob.negocio.contexto.InformacoesUsuario;

public class ContaCapitalIntegracaoInternacionalizacaoInterceptorTest {

	@BeforeClass
	public static void setUpClass() {
		InformacoesUsuario.INSTANCIA.set(new InformacoesUsuario());
	}
	
	@Test
	public void deveInstanciarInterceptor() {
		ContaCapitalIntegracaoInternacionalizacaoInterceptor interceptor = 
				new ContaCapitalIntegracaoInternacionalizacaoInterceptor();
		Assert.assertNotNull(interceptor);
		Assert.assertNotNull(interceptor.getResourceBundle());
	}
	
}
